rename files and components
